requirements.txt
fastapi
fastapi[standard]
pydantic
scikit-learn
numpy
pytest
pytest-asyncio
mlflow
python-dotenv
